contributors are benied and only allowed the creator
no one can add  or fork or bridge or open a business or bank or exchange or platform with out have authorization as a  meassge will be send to any type of the creators online. a runing test on all of the internet web 3  to identify ownership. as blockchain is the main security with a new  code over the  current code update if only if it is by creator  verification need to make any changes  or undates  or to add on all documents. All information and documents are locked and can not be unlocked by anyone but the creator.